+++ amazon_creators_async/__init__.py
@@ -1,19 +1,59 @@
 from .client import AmazonCreatorsAsyncClient
-from .exceptions import AmazonCreatorsException, RateLimitError, AuthenticationError
-from .models.requests import SearchItemsRequest, GetItemsRequest
-from .models.responses import SearchItemsResponse, GetItemsResponse
+from .exceptions import (
+    AmazonCreatorsException,
+    RateLimitError,
+    AuthenticationError,
+    InvalidRequestError,
+    APIError,
+)
+from .models.requests import (
+    GetBrowseNodesRequest,
+    GetItemsRequest,
+    GetVariationsRequest,
+    SearchItemsRequest,
+)
+from .models.responses import (
+    GetBrowseNodesResponse,
+    GetItemsResponse,
+    GetVariationsResponse,
+    SearchItemsResponse,
+)
+from .resources import (
+    GetBrowseNodesResources,
+    GetItemsResources,
+    GetVariationsResources,
+    Resources,
+    SearchItemsResources,
+)
 from .utils import Region
 
 __all__ = [
+    # Client
     "AmazonCreatorsAsyncClient",
+    # Exceptions
     "AmazonCreatorsException",
     "RateLimitError",
     "AuthenticationError",
-    "SearchItemsRequest",
+    "InvalidRequestError",
+    "APIError",
+    # Request models
+    "GetBrowseNodesRequest",
     "GetItemsRequest",
-    "SearchItemsResponse",
+    "GetVariationsRequest",
+    "SearchItemsRequest",
+    # Response models
+    "GetBrowseNodesResponse",
     "GetItemsResponse",
-    "Region"
+    "GetVariationsResponse",
+    "SearchItemsResponse",
+    # Resource constants
+    "GetBrowseNodesResources",
+    "GetItemsResources",
+    "GetVariationsResources",
+    "Resources",
+    "SearchItemsResources",
+    # Utils
+    "Region",
 ]
 
 __version__ = "0.1.2"